requirements.txt
Flask==1.1.1
PyYAML==5.1.2
uWSGI==2.0.18
